requirements.txt
Flask
gunicorn
requests
pyDes
flask-cors
python-dotenv
schedule
